+++ RELEASE_NOTES.md
@@ -37,6 +37,11 @@
 - Rename the experimental enum `EpicFightInputActions` to `EpicFightInputAction` to follow Java naming
   conventions. [#2194](https://github.com/Epic-Fight/epicfight/issues/2194)
 - Removed AirAttack and its related fields (SkillCategory, SkillSlot) to merge air slash and combo attacks as one skill
+- Updated the experimental Epic Fight's input API to support using custom input actions that are not a
+  `EpicFightInputAction`.
+  [#2194](https://github.com/Epic-Fight/epicfight/issues/2194)
+- Deprecated `ClientEngine#isBattleMode` and added `ClientEngine#isEpicFightMode` for a smoother migration when porting
+  from MC 1.20.1 to 1.21.1
 
 ## [20.13.6] - 2025-11-12